pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "allianceauth-corptools-finances"
description = "allianceauth-corptools-finances Plugin for Alliance Auth (GitHub Version)"
readme = "README.md"
keywords = [
    "allianceauth",
    "allianceauth-corptools-finances",
]
license = { file = "LICENSE" }
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.10,<3.14"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.3.1,<5",
    "allianceauth-corptools>=3.0.0b3,<4",
]
urls.Changelog = "https://github.com/ppfeufer/allianceauth-corptools-finances/blob/master/CHANGELOG.md"
urls.Tracker = "https://github.com/ppfeufer/allianceauth-corptools-finances/issues"

[tool.hatch.build]
include = [
    "/finances",
]
[tool.hatch.version]
path = "finances/__init__.py"
